web-preferences-api/contrast.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>WebPreferences Test: contrast attribute tests</title>
    <link rel="author" title="Luke Warlow" href="mailto:luke@warlow.dev" />
    <link rel="help" href="https://wicg.github.io/web-preferences-api/" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(() => {
        assert_true('preferences' in navigator, 'navigator.preferences defined');
        assert_true('contrast' in navigator.preferences, 'navigator.preferences.contrast defined');
      }, "navigator.preferences.contrast supported");
      test(() => {
        navigator.preferences.contrast = 'more';
        assert_true(window.matchMedia('(prefers-contrast: more)').matches, '(prefers-contrast: more) matches');
        navigator.preferences.contrast = 'less';
        assert_true(window.matchMedia('(prefers-contrast: less)').matches, '(prefers-contrast: less) matches');
        navigator.preferences.contrast = 'no-preference';
        assert_true(window.matchMedia('(prefers-contrast: no-preference)').matches, '(prefers-contrast: no-preference) matches');
        navigator.preferences.contrast = null;
        assert_true(window.matchMedia('(prefers-contrast: no-preference)').matches, '(prefers-contrast: no-preference) matches');
      }, "navigator.preferences.contrast updates (prefers-contrast)");
      test(() => {
        assert_throws_js(TypeError, () => { navigator.preferences.contrast = 'custom'; });
        assert_false(window.matchMedia('(prefers-contrast: custom)').matches, '(prefers-contrast: custom) doesn\'t match');
        navigator.preferences.contrast = null;
      }, "navigator.preferences.contrast = 'custom' doesn't updates (prefers-contrast)");
      test(() => {
        assert_throws_js(TypeError, () => { navigator.preferences.contrast = 'this-is-an-invalid-value'; });
        assert_equals(null, navigator.preferences.contrast);
        navigator.preferences.contrast = null;
      }, "navigator.preferences.contrast set to invalid value throws TypeError");
    </script>
  </body>
</html>
